Rename logger api module.
